routes: Modularized routes handling.
For a small project which is not production grade, writing all the routes in the
`index.js` file wouldn't be an issue. However as your codebase increases and
the number of routes increase with time, you need to have clear idea of what does
what and hence, giving each api module its own route file becomes easier. You can
increase the depth of directory structure by having separate files for separate
methods of same module inside the routes folder of the directory. For example
over here we could do :-

` + routes/
   ++ movies/
       ++ getAllMovies.js
       ++ getMovieById.js 
`
Read more on `routes`, `module.exports` and `modular routes`.
